leo_typeidea/requirements.txt
Package                   Version  
------------------------- ---------
backports.csv             1.0.7    
certifi                   2019.6.16
chardet                   3.0.4    
defusedxml                0.6.0    
diff-match-patch          20181111 
Django                    1.11.24  
django-autocomplete-light 3.2.10   
django-ckeditor           5.4.0    
django-crispy-forms       1.7.2    
django-formtools          2.1      
django-import-export      1.2.0    
django-js-asset           1.2.2    
django-rest-framework     0.1.0    
django-reversion          3.0.4    
djangorestframework       3.10.3   
et-xmlfile                1.0.1    
future                    0.17.1   
httplib2                  0.9.2    
idna                      2.8      
jdcal                     1.4.1    
Markdown                  3.1.1    
mistune                   0.8.4    
odfpy                     1.4.0    
openpyxl                  2.6.3    
pbr                       5.4.3    
Pillow                    5.1.0    
pip                       19.0.3   
pytz                      2019.2   
PyYAML                    5.1.2    
requests                  2.22.0   
setuptools                40.8.0   
six                       1.12.0   
stevedore                 1.31.0   
tablib                    0.13.0   
urllib3                   1.25.3   
virtualenv                16.7.5   
virtualenv-clone          0.5.3    
virtualenvwrapper         4.8.4    
xadmin                    0.6.1    
xlrd                      1.2.0    
xlwt                      1.3.0    
